agents/aligner/requirements.txt
pydub
ffmpeg-python
python-dotenv
